pyproject.toml
[tool.poetry]
name = "aws-sso-util"
version = "1.4.0"
description = "Utilities to make AWS SSO easier"
authors = ["Ben Kehoe <ben@kehoe.io>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/benkehoe/aws-sso-util"

[tool.poetry.scripts]
aws-sso-util = 'aws_sso_util.cli:cli'

[tool.poetry.dependencies]
python = "^3.6"
# botocore = {git = "https://github.com/boto/botocore.git", rev = "v2"}
click = "^7.1.2"
boto3 = "^1.16.13"
pyyaml = "^5.3.1"
jsonschema = "^3.2.0"
aws-error-utils = "^1.0.4"

[tool.poetry.dev-dependencies]
pylint = "^2.5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
